old compiler and os fix

+++ test/quic_multistream_test.c
@@ -1989,9 +1989,9 @@ out:
 
                 TEST_info("client side is closed: %llu(%s)/%llu(%s), "
                           "%s, %s, reason: \"%s\"",
-                    (unsigned long long)cc_info.error_code,
+                    (uint64_t)cc_info.error_code,
                     e_str,
-                    (unsigned long long)cc_info.frame_type,
+                    (uint64_t)cc_info.frame_type,
                     f_str,
                     (cc_info.flags & SSL_CONN_CLOSE_FLAG_LOCAL) != 0
                         ? "local"
@@ -2017,9 +2017,9 @@ out:
 
             TEST_info("server side is closed: %llu(%s)/%llu(%s), "
                       "%s, %s, reason: \"%s\"",
-                (unsigned long long)tcause->error_code,
+                (uint64_t)tcause->error_code,
                 e_str,
-                (unsigned long long)tcause->frame_type,
+                (uint64_t)tcause->frame_type,
                 f_str,
                 tcause->remote ? "remote" : "local",
                 tcause->app ? "app" : "transport",
